requirements.txt
python-dotenv==1.0.1
dbt-core==1.8.4
dbt-postgres==1.8.2
pandas==2.2.2